requirements.txt
distlib==0.3.6
filelock==3.9.0
mysql-connector-python==8.0.31
platformdirs==2.6.2
protobuf==3.20.1
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
virtualenv==20.17.1